Moved Starter and Client to different packages to guarantee ordering for component scanning.
Previously the starter and the client (app) where living in the same package “demo” which didn’t support to redefine a “HelloService” bean with a different name than “helloService” via an @Bean method in App.

In that case one had to either define a bean with the same name as the bean from the auto-configuration or move the new bean definition to a different package (or nested static class).